requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
jsonify==0.5
jwt==1.3.1
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
requests==2.26.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
email-validator==1.1.3
PyJWT==2.3.0
python-decouple==3.5